SAFRAN-537: SQL sequence import can handle BigInteger
The import of sequences with the different connectors failed if the
values of the elements imported were bigger than the size of an int
(2147483647). 

The connectors can now handle BigInteger size and even if this size is
exceeded, the value will be null but the import can continue.

+++ reversers/database/plugins/org.obeonetwork.dsl.database.reverse/src/org/obeonetwork/dsl/database/reverse/utils/CreationUtils.java
@@ -1,5 +1,7 @@
 package org.obeonetwork.dsl.database.reverse.utils;
 
+import java.math.BigInteger;
+
 import org.obeonetwork.dsl.database.AbstractTable;
 import org.obeonetwork.dsl.database.Column;
 import org.obeonetwork.dsl.database.Constraint;
@@ -101,7 +103,7 @@ public class CreationUtils {
 		return constraint;
 	}
 	
-	public static Sequence createSequence(TableContainer owner, String name, Integer increment, Integer minValue, Integer maxValue, Integer start, boolean cycle, Integer cache) {
+	public static Sequence createSequence(TableContainer owner, String name, BigInteger increment, BigInteger minValue, BigInteger maxValue, BigInteger start, boolean cycle, BigInteger cache) {
 		Sequence sequence = DatabaseFactory.eINSTANCE.createSequence();
 		sequence.setName(name);
 		sequence.setIncrement(increment);
